+++ HISTORY.md
@@ -2,6 +2,9 @@
 
 ## Unrelease
 
+1. 增加 `--data-validator-no-error-version` ，仅仅低于这个版本的验证器会报错。高于或等于这个版本的验证器仅warning。用于新增不兼容规则时，预留版本迁移期。
+2. 自定义验证器增加 `version` 字段用于标识版本号。
+
 ### 2.20.0
 
 1. Ue-Json和Ue-Csv模式下输出的代码中转出Enum类型
@@ -19,7 +22,6 @@
 ## 2.19.2
 
 1. 修复 `--data-source-mapping-file` 指向的地址无效时的 stacktrace 。
-
 
 ## 2.19.1